add older scripts
They were created a while ago. I'm adding them in case I need them later but I will be deletin gost of them in the next comit

A lot of the script were from either small tasks I wanted to perform with the data or data validation which I tried to make work for a while but it was too much to review and not very precise.